requirements.txt
streamlit 
google.generativeas
pypdf
opencv-python
python-docx
pillow